requirements.txt
absl-py==0.9.0
astor==0.8.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
numpy==1.18.2
opt-einsum==3.2.0
protobuf==3.11.3
PyYAML==5.3.1
scipy==1.4.1
six==1.14.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu=1.15.2
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
